Split rendering and storybook setup of Epic into two files (along lines of other components)
