+++ src/BrazeMessage.tsx
@@ -35,7 +35,6 @@ const COMPONENT_MAPPINGS: ComponentMapping = {
     [APP_BANNER_NAME]: AppBanner,
     [SPECIAL_EDITION_BANNER_NAME]: SpecialEditionBanner,
     [THE_GUARDIAN_IN_2020_BANNER_NAME]: TheGuardianIn2020Banner,
-    [EPIC_NAME]: Epic,
 };
 
 export type Props = {
